refactor: Use existing Program and Startup for samples project, with target tweaks
