+++ samples/Samples.Server/Startup.cs
@@ -10,22 +10,47 @@ using Microsoft.Extensions.DependencyInjection;
 using System;
 using System.Collections.Generic;
 
+#if NETCOREAPP2_2
+#else
+using Microsoft.AspNetCore.Server.Kestrel.Core;
+using Microsoft.Extensions.Hosting;
+#endif
+
 namespace GraphQL.Samples.Server
 {
     public class Startup
     {
+#if NETCOREAPP2_2
         public Startup(IConfiguration configuration, IHostingEnvironment environment)
+#else
+        public Startup(IConfiguration configuration, IWebHostEnvironment environment)
+#endif
         {
             Configuration = configuration;
             Environment = environment;
         }
 
         public IConfiguration Configuration { get; }
+
+#if NETCOREAPP2_2
         public IHostingEnvironment Environment { get; }
+#else
+        public IWebHostEnvironment Environment { get; }
+#endif
 
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
+#if NETCOREAPP3_0
+            // Workaround until GraphQL can swap off Newtonsoft.Json and onto the new MS one.
+            // Depending on whether you're using IIS or Kestrel, the code required is different
+            // See: https://github.com/graphql-dotnet/graphql-dotnet/issues/1116
+            services.Configure<KestrelServerOptions>(options =>
+            {
+                options.AllowSynchronousIO = true;
+            });
+#endif
+
             services
                 .AddSingleton<IChat, Chat>()
                 .AddSingleton<ChatSchema>()
